+++ test/looking-glass-test.el
@@ -1,8 +1,6 @@
 ;;; looking-glass-test.el --- Test entrypoint for looking-glass -*- lexical-binding: t; -*-
 
 (require 'looking-glass-core-test)
-(require 'looking-glass-convert-test)
-(require 'looking-glass-json-test)
 (require 'looking-glass-buffer-test)
 
 ;;; looking-glass-test.el ends here